requirements.txt
tweepy>=4.14.0,<5
pytest>=7,<9
Pillow>=10.0.0,<11
